Move some long strings around
