+++ src/olmo_eval/harness/presets.py
@@ -14,7 +14,11 @@ from olmo_eval.common.types import ProviderKind
 from olmo_eval.harness.sandbox import Capability
 
 from .config import HarnessConfig, ProviderConfig
-from .constants import DR_TULU_SYSTEM_PROMPT
+from .constants import (
+    CODE_COMPLETION_SYSTEM_PROMPT,
+    CODING_AGENT_SYSTEM_PROMPT,
+    DR_TULU_SYSTEM_PROMPT,
+)
 
 
 def _get_logs_dir() -> str:
@@ -48,71 +52,6 @@ class _Lazy:
 def lazy(fn: Callable[[str], HarnessConfig]) -> _Lazy:
     """Mark a preset factory for lazy loading. Factory receives preset name."""
     return _Lazy(fn)
-
-
-# ─────────────────────────────────────────────────────────
-# System Prompts
-# ─────────────────────────────────────────────────────────
-
-CODING_AGENT_SYSTEM_PROMPT = """\
-You are a helpful coding assistant with access to a sandboxed bash shell.
-
-You can execute bash commands to:
-- Run code and tests
-- Install packages
-- Manipulate files
-- Explore the filesystem
-
-You can also search the web for documentation and examples using the provided tools.
-
-Use the execute_bash tool to run commands. The environment is isolated,
-so you can safely experiment.
-
-IMPORTANT: When writing code with special characters (quotes, backslashes, newlines),
-use a heredoc to write the code to a file, then run the file:
-
-cat << 'EOF' > solution.py
-# Your code here - special characters are preserved exactly
-def example():
-    return "hello\\nworld"
-EOF
-python solution.py
-
-Do NOT use python -c or inline code with lots of special characters.
-
-When solving coding problems:
-1. First understand the problem by reading any provided files
-2. Plan your approach and write code in a file using heredoc syntax
-3. Run and test your solution
-4. Verify it works before providing the final answer
-"""
-
-CODE_COMPLETION_SYSTEM_PROMPT = """\
-You are a Python coding assistant that completes function implementations.
-
-When given a function signature and docstring, write the implementation code that \
-goes inside the function body. Output only valid Python code.
-
-You have access to tools to help you:
-- execute_bash: Run Python code in a sandbox to test your solution
-- Web search: Look up documentation or examples if needed
-
-Workflow:
-1. Read the function signature and docstring carefully
-2. Write the implementation code
-3. Test your code using execute_bash to verify it works
-4. Provide the final implementation
-
-When testing code, write it to a file using heredoc syntax:
-
-cat << 'EOF' > solution.py
-# Your implementation here
-EOF
-python solution.py
-
-Output only the function body code in your final answer - no explanations, \
-markdown formatting, or the function signature itself.
-"""
 
 
 # ─────────────────────────────────────────────────────────
